split mitmproxy.flow into submodules
