+++ mitmproxy/flow/compat.py
@@ -2,7 +2,7 @@
 This module handles the import of mitmproxy flows generated by old versions.
 """
 from __future__ import absolute_import, print_function, division
-from . import version
+from .. import version
 
 
 def convert_013_014(data):
