+++ mitmproxy/flow/export.py
@@ -7,6 +7,7 @@ from netlib.utils import parse_content_type
 import re
 from six.moves.urllib.parse import urlparse, quote, quote_plus
 
+
 def curl_command(flow):
     data = "curl "
 
@@ -123,18 +124,18 @@ def locust_code(flow):
             max_wait = 3000
 """).strip()
 
-
     components = map(lambda x: quote(x, safe=""), flow.request.path_components)
     file_name = "_".join(components)
     name = re.sub('\W|^(?=\d)', '_', file_name)
     url = flow.request.scheme + "://" + flow.request.host + "/" + "/".join(components)
     if name == "" or name is None:
-        new_name = "_".join([str(flow.request.host) , str(flow.request.timestamp_start)])
+        new_name = "_".join([str(flow.request.host), str(flow.request.timestamp_start)])
         name = re.sub('\W|^(?=\d)', '_', new_name)
     args = ""
     headers = ""
     if flow.request.headers:
-        lines = ["            '%s': '%s',\n" % (k, v) for k, v in flow.request.headers.fields if k.lower() not in ["host", "cookie"]]
+        lines = ["            '%s': '%s',\n" % (k, v) for k, v in flow.request.headers.fields if
+                 k.lower() not in ["host", "cookie"]]
         headers += "\n        headers = {\n%s        }\n" % "".join(lines)
         args += "\n            headers=headers,"
