+++ mitmproxy/onboarding/__init__.py
@@ -1,0 +1,5 @@
+from .cert_app import app
+
+__all__ = [
+    "app"
+]
